Move EFW e2e tests into its own file
